requirements.txt
aiofiles==24.1.0
anyio==4.10.0
beautifulsoup4==4.13.5
build==1.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
Cython==3.1.3
eventkit==1.0.3
exchange_calendars==4.11.1
finnhub-python==2.4.24
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
ib-insync==0.9.86
idna==3.10
korean-lunar-calendar==0.3.1
llvmlite==0.44.0
markdown-it-py==4.0.0
mdurl==0.1.2
nest-asyncio==1.6.0
numba==0.61.2
numpy==2.2.6
packaging==25.0
pandas==2.3.2
pandas_market_calendars==5.1.1
Pygments==2.19.2
pyluach==2.3.0
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
rich==14.1.0
ruff==0.12.12
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
TA-Lib==0.6.7
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
